HADOOP-19282. STSClientFactory: do not use URIBuilder (#7966) (#8008)
URIBuilder was used from the AWS SDK for Java v2, from the
shaded Apache HTTP Client. 

It is a problem if a user would like not to
use the AWS SDK bundle, since more or less only 3 modules are needed
(s3, s3-transfer & sts), but that may cause problems on unshaded
dependency versions. Since a URI constructor can achieve the same here I
switched it as a preferred option.

Contributed by Lapshin Dmitry