HADOOP-19696. hadoop binary distribution to move cloud connectors to hadoop common/lib (#7980) (#8094)
This moves all the cloud connector libraries to common/lib
There are specific build options to control which libraries to include
The hadoop-* JARs of the modules are includes, but dependencies
are only included when the build-time options specify it.

  Available package profiles:
    hadoop-aliyun-package
    hadoop-aws-package
    hadoop-azure-datalake-package
    hadoop-cos-package
    hadoop-huaweicloud-package

This means that by default AWS bundle.jar is no longer included
in the distribution: to add it users must drop their chosen version
of the SDK into share/hadoop/common/lib

Anyone building their own release now has a choice of which connectors
to bundle. The ASF ones will stay fairly lean to reduce the CVE
attack surface as well as keep package size under control.

+ followups
* HADOOP-19696. restore hadoop-aliyun-package to docs
* HADOOP-19696. Followup: Li2cense spelling in assembly xml files

Contributed by Steve Loughran